[broken] fix a few warnings

+++ src/vmath.h
@@ -14,11 +14,11 @@
 
 namespace
 {
-    constexpr const float DegPerRad = 180 / M_PI;
-    constexpr const float RadPerDeg = M_PI / 180;
-    constexpr const float Rad90 = 0.5*M_PI;
-    constexpr const float Rad180 = M_PI;
-    constexpr const float Rad360 = 2 * M_PI;
+    constexpr const float DegPerRad = static_cast<float>(180 / M_PI);
+    constexpr const float RadPerDeg = static_cast<float>(M_PI / 180);
+    constexpr const float Rad90 = static_cast<float>(0.5*M_PI);
+    constexpr const float Rad180 = static_cast<float>(M_PI);
+    constexpr const float Rad360 = static_cast<float>(2 * M_PI);
 }
 
 /**
